+++ lib/rspec-api/matchers/attributes.rb
@@ -1,5 +1,5 @@
 RSpec::Matchers.define :have_attribute do |name, options = {}|
-  name, can_be_nil, type = name.to_s, options[:can_be_nil], options[:type]
+  name, can_be_nil, type, format = name.to_s, options[:can_be_nil], options[:type], options[:format]
 
   match do |json|
     Array.wrap(json).all? do |item|
@@ -8,7 +8,7 @@ RSpec::Matchers.define :have_attribute do |name, options = {}|
       elsif can_be_nil
         item.key?(name)
       else
-        matches_type?(item[name], type)
+        matches_type?(item[name], type) && matches_format?(item[name], format)
       end
     end
   end
@@ -23,12 +23,22 @@ RSpec::Matchers.define :have_attribute do |name, options = {}|
   end
 end
 
-def matches_type?(value, type)
+def matches_type?(value, type) # JSON => Ruby type conversion
   case type
+    when :number then value.is_a? Numeric # TODO: Double-precision floating-point
+    when :array then value.is_a? Array
+    when :object then value.is_a? Hash
+    when :string then value.is_a? String
+    when :null then value.nil?
+    when :boolean then [TrueClass, FalseClass].include? value.class
+  end
+end
+
+def matches_format?(value, format)
+  case format
     when :url then value =~ URI::regexp
     when :timestamp then DateTime.iso8601 value rescue false
-    when :boolean then [TrueClass, FalseClass].include? value.class
     when :email then value =~ %r{(?<name>.+?)@(?<host>.+?)\.(?<domain>.+?)}
-    else value.is_a? type.to_s.classify.constantize
+    else true
   end
 end